chore: changing the folder structure of the unity and tests datasources folders
