+++ src/graphql/datasources/news-api/tests/unity/news-api.unity.spec.ts
@@ -1,11 +1,14 @@
 const mockRestDataSourceGet = jest.fn();
 
-import { rawArticleWithId, articleWithId } from '../../../../__tests__/mocks/articles';
-import { ArticleLanguage } from '../../../lib/types';
-import env from '../../../config/environment';
-import { makeDateParam } from './helpers';
-import CONSTANTS from './utils/constants';
-import NewsAPI from './NewsAPI';
+import { ArticleLanguage } from '../../../../../lib/types';
+import {
+  rawArticleWithId,
+  articleWithId,
+} from '../../../../../../__tests__/mocks/articles';
+import env from '../../../../../config/environment';
+import { makeDateParam } from '../../helpers';
+import CONSTANTS from '../../utils/constants';
+import NewsAPI from '../../NewsAPI';
 
 const dateParam = makeDateParam();
