--- src/graphql/datasources/the-movie-db-api/tests/integration/makeTestQuery.ts
@@ -1,22 +1,0 @@
-import { createTestClient } from 'apollo-server-testing';
-import { ApolloServer } from 'apollo-server';
-
-import resolvers from '../../../../resolvers';
-import typeDefs from '../../../../typeDefs';
-import TheMovieDBAPI from '../../TheMovieDBAPI';
-
-const makeTestQuery = () => {
-  const server = new ApolloServer({
-    dataSources: () => ({
-      tmdb: new TheMovieDBAPI(),
-    }),
-    resolvers,
-    typeDefs,
-  });
-
-  const { query } = createTestClient(server);
-
-  return query;
-};
-
-export default makeTestQuery;
